jobs/send_bday_notification/requirements.txt
color-it
pendulum
mysql-connector-python
requests
python-dotenv